+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Added new commands `delete consumer-group` to delete a consumer-group
+
 ## 1.18.1 - 2021-07-15
 
 ### Fixed